pyproject.toml
[project]
name = "dropdock"
version = "0.1.0"
description = "A CLI to automatically pull and manage Spamhaus DROP lists into PiHole."
authors = [
    {name = "Jake Lindsay",email = "jake@jakelindsay.uk"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "pydantic (>=2.11.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project.scripts]
dockdrop = "dropdock.cli:main"